requirements.txt
beautifulsoup4==4.12.3
google_api_python_client==2.153.0
isodate==0.7.2
pytz==2024.2
Requests==2.32.3
python-dotenv==1.0.1